Refactored config variables to base, oauth & unique
