+++ tests/functional/config_files/patientAccessCfg.py
@@ -1,30 +1,20 @@
-from ...scripts.enviroment import EnvVarWrapper
+from ...scripts.environment import EnvVarWrapper
 
 
 ENV = EnvVarWrapper(
     **{
         "id_token_nhs_login_private_key_absolute_path": "ID_TOKEN_NHS_LOGIN_PRIVATE_KEY_ABSOLUTE_PATH",
-        "test_patient_id": "TEST_PATIENT_ID",
-        "oauth_proxy": "OAUTH_PROXY",
-        "oauth_base_uri": "OAUTH_BASE_URI",
-        "environment": "APIGEE_ENVIRONMENT",
         "jwt_private_key_absolute_path": "JWT_PRIVATE_KEY_ABSOLUTE_PATH",
-        "pds_base_path": "PDS_BASE_PATH",
         "client_id": "CLIENT_ID",
         "client_secret": "CLIENT_SECRET",
-        "redirect_uri": "REDIRECT_URI"
+        "redirect_uri": "REDIRECT_URI",
     }
 )
 
 CONFIG = {
     "ID_TOKEN_NHS_LOGIN_PRIVATE_KEY_ABSOLUTE_PATH": ENV["id_token_nhs_login_private_key_absolute_path"],
-    "TEST_PATIENT_ID": ENV['test_patient_id'],
-    "OAUTH_PROXY": ENV["oauth_proxy"],
-    "OAUTH_BASE_URI": ENV["oauth_base_uri"],
-    "BASE_URL": f"https://{ENV['environment']}.api.service.nhs.uk",
     "JWT_PRIVATE_KEY_ABSOLUTE_PATH": ENV["jwt_private_key_absolute_path"],
-    "PDS_BASE_PATH": ENV["pds_base_path"],
     "CLIENT_ID": ENV['client_id'],
     "CLIENT_SECRET": ENV['client_secret'],
-    "REDIRECT_URI": ENV['redirect_uri']
+    "REDIRECT_URI": ENV['redirect_uri'],
 }
